+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.7.0](https://github.com/0xC0000054/gmic-sharp-native/compare/v0.6.0...v0.7.0) - 2020-07-10
+
 ### Added
 
 * Allow callers to set a host application name.
@@ -45,4 +47,3 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 First version
-